refactored text processing functions into a TextProcessor object.
